rename rpc variables to dispatcher
