+++ examples/thrift/keyvalue/client/main.go
@@ -69,7 +69,7 @@ func main() {
 		Filter:    cache,
 	})
 	if err := dispatcher.Start(); err != nil {
-		log.Fatalf("failed to start RPC: %v", err)
+		log.Fatalf("failed to start Dispatcher: %v", err)
 	}
 	defer dispatcher.Stop()
